+++ app/src/main/java/com/datn/bia/a/domain/model/dto/res/ResProductDTO.kt
@@ -12,7 +12,7 @@ data class ResProductDataDTO(
     @SerializedName("_id")
     val id: String? = null,
     val name: String? = null,
-    val price: Int? = null,
+    val price: Double? = null,
     val variants: List<ResVariantDTO>? = null,
     @SerializedName("caterori")
     val category: ResCatProductDTO? = null,
@@ -35,7 +35,7 @@ data class ResCatProductDTO(
 
 data class ResVariantDTO(
     val color: String? = null,
-    val price: Int? = null,
+    val price: Double? = null,
     val quantity: Int? = null,
     val status: Boolean? = null,
     val _id: String? = null
