+++ app/src/main/java/com/datn/bia/a/domain/model/entity/ProductEntity.kt
@@ -1,5 +1,6 @@
 package com.datn.bia.a.domain.model.entity
 
+import android.renderscript.Double3
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.datn.bia.a.domain.model.dto.res.ResVariantDTO
@@ -9,7 +10,7 @@ data class ProductEntity(
     @PrimaryKey(autoGenerate = false)
     val id: String,
     val name: String,
-    val price: Int,
+    val price: Double,
     val imageUrl: String,
     val albumImage: List<String>,
     val des: String,
